add: WeakCallback 弱回调实现

+++ utility/Types.h
@@ -4,22 +4,22 @@
 
 #ifndef CPP_TYPES_H
 #define CPP_TYPES_H
+
 #include <stdint.h>
 #include <string.h>  // memset
 #include <string>
 
 #ifndef NDEBUG
+
 #include <assert.h>
+
 #endif
 
-namespace zhengqi
-{
-    namespace utility
-    {
+namespace zhengqi {
+    namespace utility {
         using std::string;
 
-        inline void memZero(void* p, size_t n)
-        {
+        inline void memZero(void *p, size_t n) {
             memset(p, 0, n);
         }
 
@@ -78,9 +78,8 @@ namespace zhengqi
 // but the proposal was submitted too late.  It will probably make
 // its way into the language in the future.
         template<typename To, typename From>
-        inline To implicit_cast(From const &f)
-        {
-            return f;
+        inline To implicit_cast(From const &f) {
+            return (unsigned long) f;
         }
 
 // When you upcast (that is, cast a pointer from type Foo to type
@@ -101,16 +100,16 @@ namespace zhengqi
 //    if (dynamic_cast<Subclass2>(foo)) HandleASubclass2Object(foo);
 // You should design the code some other way not to need this.
 
-        template<typename To, typename From>     // use like this: down_cast<T*>(foo);
-        inline To down_cast(From* f)                     // so we only accept pointers
+        template<typename To, typename From>
+        // use like this: down_cast<T*>(foo);
+        inline To down_cast(From *f)                     // so we only accept pointers
         {
             // Ensures that To is a sub-type of From *.  This boost_test is here only
             // for compile-time type checking, and has no overhead in an
             // optimized build at run-time, as it will be optimized away
             // completely.
-            if (false)
-            {
-                implicit_cast<From*, To>(0);
+            if (false) {
+                implicit_cast<From *, To>(0);
             }
 
 #if !defined(NDEBUG) && !defined(GOOGLE_PROTOBUF_NO_RTTI)
